PoC Sensitivity analysis tabs
We now have a different view with tabs as a PoC for the sensitivity analysis. Only problem is that the white bow behind the output don't work anymore. I know which line to comment to make them appear again (the clear_ouputs() one at the start of the display, but if I do this, menu superimpose which is not pretty. Need to think of something to solve that