PoC for automated scan of process
Added the bare bones of a feature that would allow to automatically scan existing results to prepare for post processing. This is done by scanning when there is a tab change where the old tab was the launch tab. For now it only prints a "Plop". A dummy tab was also added for PoC purposes