Added wing geometry tab
Added a tab to display the wing geometry, it suffers the same problem as the previous tab in that when something is displayed, you need to scroll down to have the full picture but it isn't much of an issue.
Also I opted out of using SelectMultiple widget because it does not drop down. instead I use a Dropdown that add stuff to the graph as you select more.
Finally, added an info tooltip to make it clear how it works