Added a payload range diagram tab
Added a payload range diagram tab base on Breguet's range equation and a corrective coefficient for range. There are no way to analytically compute the range so were using a solver from scipy. Also this seems to be the tab too much as now the tabs are wider than the picture.