Change in reference AC
Changed the reference AC to something further to an A320 as the goal of the second exercise will be to find the right data for an A320. This however, showed that for a Mach number of 0.8 the code would not converge, which made me implement a check to rearrange the sweep in the backend in case the M is too high. We also display a warning when this is the case