Removed CSS Background
The custom CSS background was removed because it paused too many compatibility issues with Voila. The original workaround I had proved not robust enough so we'll opt for a more sober design but one that allows to use voila. Added the logos as ipywidgets Images which prompted the birth of common layout for the bottom layer of each menu with 3 boxes (40% left aligned, 20% centered, 40% right aligned with logos).