MDO mode implementation
The MDO mode is now almost fully implemented. The cruise mach was changed for the sweep angle as COBYLA has a tendency to go out of bounds which will be problematic for a Mach that can't go above 1 or a math domain error appears. Now, we will try everything out.