Display objective evolution
Added a display to see the evolution of the objective of the optimization as a number of function call. Now when launching an MDA/MDO both display clear and when the MDA/MDO finishes, the according display is puth forth.
Also solved a bug that resizes the display when checking the MDO box by simply resizing them and refreshing the display